sha : update sha_binary_operation prototype to 64bits

+++ includes/ft_ssl.h
@@ -56,10 +56,10 @@ typedef struct				s_ssl
 
 typedef struct				s_md5
 {
-	uint32_t				A;
-	uint32_t				B;
-	uint32_t				C;
-	uint32_t				D;
+	uint64_t				A;
+	uint64_t				B;
+	uint64_t				C;
+	uint64_t				D;
 	size_t					dft_size;
 	size_t					pad_size;
 
@@ -116,20 +116,21 @@ t_sha512   		*init_new_sha512(size_t input_size);
 void 			sha512_processing(t_sha512 *sha512, t_ssl *ssl);
 void   			clear_sha512(t_sha512 *sha512);
 
-/*		SHA256 BINARY OPERATIONS	*/
+/*		SHA BINARY OPERATIONS	*/
 
-uint32_t	rightrotate(uint32_t x, uint32_t y);
-uint32_t    rightshift(uint32_t x, uint32_t y);
-uint32_t    ch(uint32_t x, uint32_t y, uint32_t z);
-uint32_t    maj(uint32_t x, uint32_t y, uint32_t z);
-uint32_t    ssig0(uint32_t x);
-uint32_t    ssig1(uint32_t x);
-uint32_t    bsig0(uint32_t x);
-uint32_t    bsig1(uint32_t x);
-uint32_t    S0(uint32_t x);
-uint32_t    S1(uint32_t x);
-uint32_t    s0(uint32_t x);
-uint32_t    s1(uint32_t x);
+uint64_t	rightrotate(uint64_t x, uint64_t y);
+uint64_t    rightshift(uint64_t x, uint64_t y);
+uint64_t    ch(uint64_t x, uint64_t y, uint64_t z);
+uint64_t    maj(uint64_t x, uint64_t y, uint64_t z);
+uint64_t    ssig0(uint64_t x);
+uint64_t    ssig1(uint64_t x);
+uint64_t    bsig0(uint64_t x);
+uint64_t    bsig1(uint64_t x);
+uint64_t    S0(uint64_t x);
+uint64_t    S1(uint64_t x);
+uint64_t    s0(uint64_t x);
+uint64_t    s1(uint64_t x);
+
 
 /*          LIBFT       */
